requirements.txt
APScheduler==3.11.2
certifi==2026.2.25
charset-normalizer==3.4.4
dotenv==0.9.9
idna==3.11
psycopg==3.3.3
psycopg-binary==3.3.3
psycopg2-binary==2.9.11
python-dotenv==1.2.1
requests==2.32.5
tzlocal==5.3.1
urllib3==2.6.3